feat: add japanese language support to ui
To be honest, it's not a perfect translation.
This translation include a lot machine translation(google-translate),
however we checked one by one and fixed clearly strange japanese.
So all japanese translation is not fluent japanese,
but it is enough to unerstand by native japanese speakers.

We can improve japanese translation regularly, and follow upstream changes.
Please consider to add japanese translation. Thank you.

Co-authored-by: Jing Lee <jasonleenz369@gmail.com>
Co-authored-by: cvabc <cvabc@users.noreply.github.com>
Co-authored-by: Hiroyuki Ikegami <ikegam@mixallow.net>