AppCenter Tibber: Access Token will be displayed as "xxx" if set (#1098)
<!-- Suggested description for FEMS Changelog (https://portal.fenecon.de/m/changelog) -->
**Tibber token will be displayed as "xxx" if set**

<!-- Technical description of changes -->
If a access token is set in the Tibber Provider it will be displayed as "xxx" and shown directly as a clear text token and can not be accessed from the ui in clear text.
If the token wont get modified and stays at "xxx" it wont get updated and the old token remains in the provider configuration.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1098
Reviewed-by: Hueseyin Sahutoglu <hueseyin.sahutoglu@fenecon.de>
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Michael Grill <michael.grill@fenecon.de>
Co-committed-by: Michael Grill <michael.grill@fenecon.de>