GoodWe: Disable ShadowManagement (#1103)
<!-- Suggested description for FEMS Changelog (https://portal.fenecon.de/m/changelog) -->
Generelle Verbesserungen an Home 10,20 & 30

<!-- Technical description of changes -->
During the time that GoodWe's shadow management is active, we will not be able to discharge or charge the battery. During this time, the DC connection cannot work as usual.
This hasn't happened in recent months and years because we have only tried to set the register once and older GoodWe firmware versions may have had a different behaviour.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1103
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Sebastian Asen <sebastian.asen@fenecon.de>
Co-committed-by: Sebastian Asen <sebastian.asen@fenecon.de>