AppCenter Tibber: Access Token will be displayed as "xxx" if set
If a access token is set in the Tibber Provider it will be displayed as "xxx" and shown directly as a clear text token and can not be accessed from the ui in clear text.
If the token wont get modified and stays at "xxx" it wont get updated and the old token remains in the provider configuration.

Co-authored-by: Michael Grill <59126309+michaelgrill@users.noreply.github.com>
Reviewed-by: Hueseyin Sahutoglu <34771592+huseyinsaht@users.noreply.github.com>
Reviewed-by: Stefan Feilmeier <3515268+sfeilmeier@users.noreply.github.com>