GoodWe: Disable ShadowManagement
During the time that GoodWe's shadow management is active, we will not be able to discharge or charge the battery. During this time, the DC connection cannot work as usual.
This hasn't happened in recent months and years because we have only tried to set the register once and older GoodWe firmware versions may have had a different behaviour.

Reviewed-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>
Reviewed-by: Stefan Feilmeier <3515268+sfeilmeier@users.noreply.github.com>