Generate maven poms and releases within bnd workspace.
This pull request introduces several changes:
1) It tidy up bnd files across modules, setting version in one place (cnf/build.bnd).
2) The license headers and vendor declaration are also updated.
3) There are additional properties which enable deployment of released artifacts to maven repositories.

Few notes, as there is an ongoing concern with `Bundle-Version` policy taken in the project.
At the moment all bundles use `1.0.0.${tstamp}` (`1.0.0.SNAPSHOT` after update) which is, even from OSGi point of view, a moving target.

Above is not in line with tags which use completely different format (`2024.4.0`).
At the moment the release version is not propagated to gradle/bnd build making it difficult to navigate between both.
For example release tag 2024.4.0 resulted in bundle version is 1.0.0.202403311530 due to different time offsets.
It is possible to synchronize these through `Bundle-Version: ${env;VERSION;1.0.0}.SNAPSHOT` construction.
Then, release scripts and pipelines needs to be updated in order to populate build environment.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>