GoodWe 20 & 30: optionally show values per MPPT; not per String
GoodWeChargerTwoStringImpl was used to represent one string from a GoodWe MPPT tracker that is responsible for two strings (GoodWe ET-Systems).

 Possible values given by GoodWe are:
 * MPPT Current
 * MPPT Power
 * String Current
 * String Voltage

As the current values of one string are incorrect (so far DSP-Version 12, ARM-Version 27) the power cannot be calculated as expected.

A new Component GoodWeChargerMpptTwoStringImpl is mirroring the GoodWe values to get the important values from
 * MPPT Current
 * MPPT Power
 * String Voltage (One of the two strings)

Co-authored-by: Michael Grill <59126309+michaelgrill@users.noreply.github.com>
Reviewed-by: Michael Grill <59126309+michaelgrill@users.noreply.github.com>
Reviewed-by: Stefan Feilmeier <3515268+sfeilmeier@users.noreply.github.com>
Co-authored-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>
Co-committed-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>