Change phase switching logic and add further tests.
The conditions when the phase switch would happen was to static in my opinion. I hope your expected behaviour is still the same.

As mentioned the tests cannot run as you expeced as it is hardly depending on the inputs and outputs of every cycle. I added a static mehtod instead, that could be tested separately.