GoodWe: ignore impossible Battery Power values (#1207)
The GoodWe inverter provides only the dc battery power (The active power will be calculated based on the battery power and the dc-pv)
For an unknown reason, the inverter is sending one impossible power value
The result is a high peak in the consumption and active power.

**Solution:**
Values greater than the max apparent power are ignored.
The last DC-Power value or (if not defined) at least the current min/max power are used instead

**Still missing:**
What caused the incorrect value?
e.g. There was no direct correlation with BMS write values

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1207
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Sebastian Asen <sebastian.asen@fenecon.de>
Co-committed-by: Sebastian Asen <sebastian.asen@fenecon.de>