Fix: Cycle time change from wait for trigger to actual time
If the config was changed from `AbstractWorker.ALWAYS_WAIT_FOR_TRIGGER_NEXT_RUN` to a value >= 0 again the cycle thread remaind in a waiting state. The reason for that is that the thread was put in an await state as soon as `AbstractWorker.ALWAYS_WAIT_FOR_TRIGGER_NEXT_RUN` was configured.
This change allows a change at releases the thread again.

Testing: Unit tests, and the simulator app starts up and starts executing cycles