Electrical Vehicle Supply Equipment (EVSE)
This is WIP for a future replacement of the current EVCS APIs. The new namespace is "Electrical Vehicle Supply Equipment (EVSE)".

This PR provides:
- APIs for
  - Electric Vehicle
  - Charge Point
- Implementations:
  - KEBA P30/P40 via Modbus/TCP (`Evse.ChargePoint.Keba`)
  - Generic Electric Vehicle (`Evse.ElectricVehicle.Generic`)
- Controllers:
  - Single Charge-Point (`Evse.Controller.Single`)
  - Cluster of Charge-Point-Controllers (`Evse.Controller.Cluster`)
- Simple UI to control the Charge-Mode of Charge-Points

NOTE: All of this might be heavily refactored in future. Having it on develop helps us to test it easier in production. It does not affect any existing EVCS implementations.