Sum: add Power and Energy distribution channels (#2923)
This PR adds a feature to closely analyze in real-time and for historic values, how the power and energy distributed between Grid, Production, Energy-Storage-System and Consumption, i.e. how much grid-buy energy was charged to the battery; how much grid-sell energy was discharged from a battery, etc.

This PR adds new Channels to the `_sum` component, that represent power and energy distribution:
- `PRODUCTION_TO_CONSUMPTION_POWER` / `PRODUCTION_TO_CONSUMPTION_ENERGY`
- `PRODUCTION_TO_GRID_POWER` / `PRODUCTION_TO_GRID_ENERGY`
- `PRODUCTION_TO_ESS_POWER` / `PRODUCTION_TO_ESS_ENERGY`
- `GRID_TO_CONSUMPTION_POWER` / `GRID_TO_CONSUMPTION_ENERGY`
- `ESS_TO_CONSUMPTION_POWER` / `ESS_TO_CONSUMPTION_ENERGY`
- `GRID_TO_ESS_POWER` / `GRID_TO_ESS_ENERGY` / `ESS_TO_GRID_ENERGY`

The unit test in `PowerDistributionTest.java` shows some example distributions:
- Grid Sell `-1000  W`, Production `2000 W`, ESS Discharge `500 W` (Consumption is `1500 W` in this case)
  - `1000 W` Production-to-Consumption
  - `1000 W` Production-to-Grid
  - `500 W` ESS-to-Consumption

These values will be useful in a UI visualization in future. The feature has been successfully tested locally on my private system. 

One downside of the approach with `CalculateEnergyFromPower` is, that the resulting energy values might be slightly different to e.g. `CONSUMPTION_ACTIVE_ENERGY`. It's difficult to overcome this - but in reality a few `Wh` difference might not make a difference.