ModbusTcp/Rtu: reset Invalid Modbus Call error after 60 seconds
Often Modbus-Api shows a "Invalid Modbus Function Call" error because e.g. IP of a relay board was incorrectly configured. Even after we have fixed the root cause, Modbus-Api remains in a fault state. Previously, this could easily be cleared by restarting the component. Since the modified method was introduced, this is no longer the case.

- Removed debounce PROCESS_IMAGE_FAULT
- Channel PROCESS_IMAGE_FAULT is set to false if there is no error within 60 seconds

Reviewed-by: Michael Grill <59126309+michaelgrill@users.noreply.github.com>
Reviewed-by: Stefan Feilmeier <3515268+sfeilmeier@users.noreply.github.com>
Co-authored-by: Fabian Brandtner <10850256+fabian94533@users.noreply.github.com>
Co-committed-by: Fabian Brandtner <10850256+fabian94533@users.noreply.github.com>