feat: change reconnection timer implementation unix time -> cpu clock time.
Currently reconnection timer use unix timestamp and it feed from system.
But this time is not motonic increasing, So there is possibility of time
go away to far future, in this case the timer doesn't work correctly.
Especially, In case of use RTC and edge is in booting time, If System
recoginize wrong time, Sometime edge set wrong timer period and wait forever.
// Actually We face this problem recently.

This change provice to use monotonic timer it based on cpu clock to
resolve ↑ problem.
Cpu clock has no external dependencies and It suitable reconnection
timer use.

Co-author: @cvabc, claude code, chatgpt codex, github copilot