[Edge] EVCS: fix twice-plugging for KEBA in Cluster
The logic in EVCS Cluster only applied charge power after EV was in READY_FOR_CHARGING mode. It seems KEBA stations often do not switch to this state internally or the information is not sent properly via UDP. This often required plugging-unplugging-replugging the cable before a charging session would actually start.

We changed the logic to start applying charge power already in NOT_READY_FOR_CHARGING mode.

There is a little risk, that a fully loaded multi-EV charge park (with empty ESS) would still start charging and would be too slow reducing the overall power. But as this can always happen anyway (e.g. when a larger load is switched on), it is anyway handled in other parts of the code already.

Reviewed-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>