[Edge] Modbus AbstractReadTask: Enhance Parsing Response; exception catch (#2879)
* Enhance Parsing Response debug output

More generic catch block for more detailed channel assignment errors. E.g. the original code is not able to catch the following kind of errors:
[modbus1 ] ERROR [dbus.api.task.AbstractReadTask] Parsing Response failed. IllegalArgumentException: Unable to set value [-4395165653183430656] for Channel [battery0/ChrW]: Cannot convert. Long [-4395165653183430656] is not fitting in Integer range.

This kind of messages would have been very helpful for detecting channel mapping errors, but it is not catched by OpenemsException

* added TestCase to cover General Exception

Trigger IllegalArgumentException by adding a new Channel, and filling it with Integer.MAX value (not fitting in short range), catch the log and check for "IllegalArgumentException"
-> For me : easiest way to catch the logError

Co-authored-by: Christian Lehne <51822163+clehne@users.noreply.github.com>