Fix NPE if mergedResult is null (or result.length is 0)
We experienced NPE with the solver strategy of "KeepingAllNearEqual" by using multiple ESS and put them into a cluster.
When those ESS were at 0 % no solution could be found and the mergedResult becomes null.

To prevent NPE we check for the null occurence and if the result.length is greater than 0 (the last one might be not nec. but I still put this here in case it might occurr)

And return null if either is the case. Same as previously done (like checking for NaN)

After implementing this change, we did not experience the NPE anymore