Adjust large screen widget size based on nav position
Changed the size attribute for the `ion-col` element in `live.component.html` to dynamically adjust the widget size on large screens depending on the navigation service's position. This allows for a more flexible layout, accommodating different navigation positions by setting the size to 8 if the navigation is on the left and 6 otherwise (standard).