[Edge] Add MQTT bridge dummy and tests
This commit introduces a dummy implementation of the MQTT bridge (`DummyBridgeMqtt.java`) and its corresponding unit tests (`BridgeMqttImplTest.java`). Additionally, it includes package information (`package-info.java`) for proper OSGi bundle export. The dummy implementation is designed for use in unit testing by providing a mock MQTT bridge environment. This allows for testing of MQTT-related functionalities without the need for an actual MQTT broker connection. The removal of the Checkstyle build command and nature from the `.project` file suggests a simplification of the build process, possibly to streamline development or due to the Checkstyle configuration being managed differently.